+++ alembic/versions/420c02357a1b_add_created_on_field.py
@@ -15,9 +15,9 @@ import sqlalchemy as sa
 
 
 def upgrade():
-    op.add_column('badges', sa.Column('created_on', sa.DateTime,
-        nullable=False)
-    )
+    op.add_column(
+        'badges',
+        sa.Column('created_on', sa.DateTime, nullable=False))
 
 
 def downgrade():