+++ tahrir_api/model.py
@@ -22,7 +22,6 @@ from sqlalchemy.orm import (
     scoped_session,
     sessionmaker,
     relationship,
-    backref,
 )